% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gpu.R
\name{check_gpu}
\alias{check_gpu}
\title{Check if GPU is Available to PyTorch}
\usage{
check_gpu()
}
\value{
\code{TRUE} if GPU is available, \code{FALSE} otherwise.
}
\description{
Check if GPU is Available to PyTorch
}
