% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_deps.R
\name{install_deps}
\alias{install_deps}
\title{Install pathml System Dependencies}
\usage{
install_deps(env_name = "pathml-r")
}
\arguments{
\item{env_name}{(optional) name of conda environment to install
dependencies into.}
}
\description{
This function installs the system dependencies for pathml.
It is recommended to run this function before using pathml.
This function will install the system dependencies into a
conda environment called \code{pathml-r}.
}
