% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathml-package.R
\docType{package}
\name{pathml-package}
\alias{pathml}
\alias{pathml-package}
\title{pathml: A friendly toolkit for Computational Pathology}
\description{
PathML is a toolbox to facilitate machine learning workflows for high-resolution whole-slide pathology images.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marchionniLab/pathml}
  \item Report bugs at \url{https://github.com/marchionniLab/pathml/issues}
}

}
\author{
\strong{Maintainer}: Lucio Rezende Queiroz \email{luciorqueiroz@gmail.com} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID})

}
\keyword{internal}
